fix: detectTitle raises YAMLParseError
The plugin fails to detect the title of Markdown files
containing YAML Front Matter and horizontal rules.

In the detectTitle() function, the regular expression
that extracts YAML Front Matter is the longest match,
so it incorrectly extracts lines with horizontal ruled
lines (---) as being YAML Front Matter.

Fixed the regular expression from the longest match to
the shortest match.